Fix uintptr_t & uint64_t Mismatch
Within Subprocess.h, ProcessHandle was written using a uint65_t for the id. However, in BuildSystemFrontend.cpp it is storing the ID using uintptr_t, and in other files, the ID is assumed to be a pointer-sized value. So it makes sense to fix this by making the id be a uintptr_t in both places.

+++ include/llbuild/Basic/Subprocess.h
@@ -58,7 +58,7 @@ namespace llbuild {
     /// Handle used to communicate information about a launched process.
     struct ProcessHandle {
       /// Opaque ID.
-      uint64_t id;
+      uintptr_t id;
     };
 
     struct ProcessInfo {
